refactor(ors_control_app): extract lib/sanitize, lib/sql, lib/warehouse
Move 8 helpers out of server/index.ts into focused lib/ modules:
- lib/sanitize.ts: sanitizeIdentifier, sanitizeFloat, sanitizeInt,
  escapeString, getSpcsToken, toIso (timestamp normaliser)
- lib/sql.ts: snowSqlLocal, snowSqlSpcs, runSql, callProcedure,
  submitSqlAsync, cancelStatement
- lib/warehouse.ts: detectWarehouse

server/index.ts now imports these instead of redefining them.
Lines down from 3,726 -> 3,545 (-181).

Phase 1 of holistic-control-app-refactor-final, lib extraction step.
No behaviour change. tsc -p tsconfig.server.json passes cleanly.
