+++ .cortex/skills/build-routing-solution/openrouteservice_app/services/ors_control_app/server/index.ts
@@ -1,15 +1,16 @@
 import express from 'express';
 import cors from 'cors';
 import { config } from 'dotenv';
-import { execSync } from 'child_process';
-import { writeFileSync, unlinkSync, readFileSync, existsSync } from 'fs';
+import { existsSync } from 'fs';
 import { join } from 'path';
-import { tmpdir } from 'os';
 import { createStudioRouter } from './studio/routes.js';
 import { reconcileStaleJobs, setRegionActivatedHandler } from './studio/jobs.js';
 import { log, getEntries, clearEntries, getUptimeMs } from './diagnostics.js';
 import { IS_SPCS, SF_DATABASE, SF_WAREHOUSE, setWarehouse, CONN, SNOWFLAKE_HOST, DEFAULT_WAREHOUSE } from './constants.js';
 import { safeRegionIdent, orsServiceName, orsServiceFqn, isDefaultRegion, currentRegionScalar, DEFAULT_REGION_NAME } from './lib/region.js';
+import { sanitizeIdentifier, sanitizeFloat, sanitizeInt, escapeString, getSpcsToken, toIso } from './lib/sanitize.js';
+import { snowSqlLocal, snowSqlSpcs, runSql, callProcedure, submitSqlAsync, cancelStatement } from './lib/sql.js';
+import { detectWarehouse } from './lib/warehouse.js';
 
 config();
 
@@ -17,51 +18,7 @@ const app = express();
 app.use(cors());
 app.use(express.json({ limit: '10mb' }));
 
-// Normalise Snowflake-returned timestamps to ISO 8601 strings so JS Date()
-// can parse them in the browser. Used by every endpoint that returns a
-// timestamp column. Without this transform, raw Snowflake values such as
-// '2026-05-11 06:52:13.367' (no timezone) cause `new Date(s)` to return
-// 'Invalid Date' in some browsers/locales.
-function toIso(v: any): any {
-  if (v == null) return v;
-  if (v instanceof Date) return isNaN(v.getTime()) ? null : v.toISOString();
-  if (typeof v === 'object' && typeof (v as any).toISOString === 'function') {
-    try { return (v as any).toISOString(); } catch { return null; }
-  }
-  if (typeof v === 'number' && Number.isFinite(v)) {
-    return new Date(v > 1e12 ? v : v * 1000).toISOString();
-  }
-  if (typeof v === 'string') {
-    const s = v.trim();
-    if (/^-?\d+(\.\d+)?$/.test(s)) {
-      const n = Number(s);
-      if (Number.isFinite(n)) return new Date(n * 1000).toISOString();
-    }
-    const m = s.match(/^(\d{4}-\d{2}-\d{2})[T ](\d{2}:\d{2}:\d{2}(?:\.\d+)?)(Z|[+-]\d{2}:?\d{2})?$/);
-    if (m) {
-      const tz = m[3] || 'Z';
-      const d = new Date(`${m[1]}T${m[2]}${tz}`);
-      return isNaN(d.getTime()) ? s : d.toISOString();
-    }
-    const d = new Date(s);
-    return isNaN(d.getTime()) ? s : d.toISOString();
-  }
-  return v;
-}
 
-async function detectWarehouse(): Promise<void> {
-  if (SF_WAREHOUSE) return;
-  try {
-    const rows = IS_SPCS
-      ? await snowSqlSpcs('SHOW WAREHOUSES LIMIT 1')
-      : snowSqlLocal('SHOW WAREHOUSES LIMIT 1');
-    const name = (rows as any[])?.[0]?.name || (rows as any[])?.[0]?.NAME;
-    if (name) setWarehouse(name);
-    else setWarehouse(DEFAULT_WAREHOUSE);
-  } catch {
-    setWarehouse(DEFAULT_WAREHOUSE);
-  }
-}
 
 async function waitForOrsGraphReady(region: string, maxWaitSecs: number = 600): Promise<{ ready: boolean; elapsed: number; profiles: string[] }> {
   const start = Date.now();
@@ -98,114 +55,6 @@ async function waitForOrsGraphReady(region: string, maxWaitSecs: number = 600): 
     await new Promise((r) => setTimeout(r, interval));
   }
   return { ready: false, elapsed: Math.round((Date.now() - start) / 1000), profiles: [] };
-}
-
-const IDENTIFIER_RE = /^[A-Za-z][A-Za-z0-9_]{0,254}$/;
-
-function sanitizeIdentifier(val: string): string {
-  const cleaned = val.replace(/[^A-Za-z0-9_]/g, '');
-  if (!IDENTIFIER_RE.test(cleaned)) throw new Error(`Invalid identifier: ${val}`);
-  return cleaned;
-}
-
-function sanitizeFloat(val: any): number {
-  const n = Number(val);
-  if (!Number.isFinite(n)) throw new Error(`Invalid number: ${val}`);
-  return n;
-}
-
-function sanitizeInt(val: any): number {
-  const n = Math.round(Number(val));
-  if (!Number.isFinite(n) || n < 0 || n > 10000) throw new Error(`Invalid integer: ${val}`);
-  return n;
-}
-
-function escapeString(val: string): string {
-  return val.replace(/\\/g, '\\\\').replace(/'/g, "''").replace(/[\x00-\x1f]/g, '');
-}
-
-function getSpcsToken(): string {
-  return readFileSync('/snowflake/session/token', 'utf-8').trim();
-}
-
-function snowSqlLocal(sql: string, database?: string, schema?: string): any[] {
-  const tmpFile = join(tmpdir(), `ors_query_${Date.now()}.sql`);
-  const db = database || SF_DATABASE;
-  let fullSql = `ALTER SESSION SET TIMEZONE='UTC';\nALTER SESSION SET query_tag = '{"origin":"sf_sit-is-fleet","name":"oss-build-routing-solution","version":{"major":1,"minor":0},"attributes":{"is_quickstart":1,"source":"sql"}}';\nUSE WAREHOUSE ${SF_WAREHOUSE};\nUSE DATABASE ${db};\n`;
-  if (schema) fullSql += `USE SCHEMA ${schema};\n`;
-  fullSql += `${sql};`;
-  writeFileSync(tmpFile, fullSql);
-  try {
-    const result = execSync(`snow sql -c ${CONN} -f "${tmpFile}" --format json 2>/dev/null`, {
-      maxBuffer: 50 * 1024 * 1024, timeout: 120000, encoding: 'utf-8',
-    });
-    const parsed = JSON.parse(result.trim());
-    if (Array.isArray(parsed) && Array.isArray(parsed[0])) return parsed[parsed.length - 1];
-    return parsed;
-  } finally {
-    try { unlinkSync(tmpFile); } catch {}
-  }
-}
-
-async function snowSqlSpcs(sql: string, database?: string, schema?: string, timeoutSecs: number = 600): Promise<any[]> {
-  const token = getSpcsToken();
-  const QUERY_TAG = '{"origin":"sf_sit-is-fleet","name":"oss-build-routing-solution","version":{"major":1,"minor":0},"attributes":{"is_quickstart":1,"source":"sql"}}';
-  const body = { statement: sql, timeout: timeoutSecs, database: database || SF_DATABASE, schema: schema || 'CORE', warehouse: SF_WAREHOUSE, parameters: { QUERY_TAG, TIMEZONE: 'UTC' } };
-  const headers: Record<string, string> = {
-    'Authorization': `Bearer ${token}`, 'Content-Type': 'application/json',
-    'Accept': 'application/json', 'X-Snowflake-Authorization-Token-Type': 'OAUTH',
-  };
-  console.log(`[SQL API] Executing: ${sql.slice(0, 200)} (WH: ${SF_WAREHOUSE}, DB: ${SF_DATABASE}, HOST: ${SNOWFLAKE_HOST})`);
-  const sqlStart = Date.now();
-  const res = await fetch(`https://${SNOWFLAKE_HOST}/api/v2/statements`, { method: 'POST', headers, body: JSON.stringify(body) });
-  if (!res.ok) {
-    const errBody = (await res.text()).slice(0, 500);
-    log('ERROR', 'SQL', `API error ${res.status}: ${errBody.slice(0, 200)}`, { durationMs: Date.now() - sqlStart });
-    throw new Error(`SQL API error ${res.status}: ${errBody}`);
-  }
-  let result: any = await res.json();
-  if (result.statementStatusUrl && (!result.data || result.code === '333334')) {
-    const pollUrl = `https://${SNOWFLAKE_HOST}${result.statementStatusUrl}`;
-    for (let i = 0; i < 120; i++) {
-      await new Promise((r) => setTimeout(r, 5000));
-      const pr = await fetch(pollUrl, { headers });
-      result = await pr.json();
-      if (result.data || (result.code && result.code !== '333334')) break;
-    }
-  }
-  if (result.message && !result.data) {
-    log('ERROR', 'SQL', `Statement error: ${result.message?.slice(0, 200)}`, { durationMs: Date.now() - sqlStart });
-    throw new Error(`SQL error: ${result.message}`);
-  }
-  if (!result.data) return [];
-  const cols = (result.resultSetMetaData?.rowType || []).map((c: any) => c.name);
-  let allData: any[][] = [...(result.data || [])];
-  const partitions = result.resultSetMetaData?.partitionInfo;
-  if (partitions && partitions.length > 1) {
-    const handle = result.statementHandle;
-    console.log(`[SQL API] Result has ${partitions.length} partitions (${result.resultSetMetaData?.numRows} rows). Fetching remaining...`);
-    for (let p = 1; p < partitions.length; p++) {
-      const pr = await fetch(`https://${SNOWFLAKE_HOST}/api/v2/statements/${handle}?partition=${p}`, { headers });
-      const partResult: any = await pr.json();
-      if (partResult.data) allData = allData.concat(partResult.data);
-    }
-  }
-  console.log(`[SQL API] Returning ${allData.length} rows`);
-  return allData.map((row: any[]) => {
-    const obj: Record<string, any> = {};
-    cols.forEach((c: string, i: number) => { obj[c] = row[i]; });
-    return obj;
-  });
-}
-
-async function runSql(sql: string, database?: string, schema?: string): Promise<any[]> {
-  if (IS_SPCS) return snowSqlSpcs(sql, database, schema);
-  return snowSqlLocal(sql, database, schema);
-}
-
-async function callProcedure(proc: string): Promise<string> {
-  const rows = await runSql(`CALL ${SF_DATABASE}.CORE.${proc}`);
-  return rows?.[0]?.[Object.keys(rows[0])[0]] || '';
 }
 
 // Idempotent boot-time init for Backload Matching (BACKLOAD_MATCHING schema +
@@ -1834,36 +1683,6 @@ app.get('/api/matrix/regions', async (_req, res) => {
     res.json({ regions: [], error: err.message });
   }
 });
-
-async function submitSqlAsync(sql: string): Promise<string> {
-  if (!IS_SPCS) {
-    runSql(sql).catch((e: any) => console.error(`[Async local] Error: ${e.message}`));
-    return `local_${Date.now()}`;
-  }
-  const token = getSpcsToken();
-  const body = { statement: sql, timeout: 0, database: SF_DATABASE, warehouse: SF_WAREHOUSE };
-  const headers: Record<string, string> = {
-    'Authorization': `Bearer ${token}`, 'Content-Type': 'application/json',
-    'Accept': 'application/json', 'X-Snowflake-Authorization-Token-Type': 'OAUTH',
-  };
-  console.log(`[SQL API Async] Submitting: ${sql.slice(0, 200)}`);
-  const r = await fetch(`https://${SNOWFLAKE_HOST}/api/v2/statements`, { method: 'POST', headers, body: JSON.stringify(body) });
-  const result = await r.json();
-  return result.statementHandle || '';
-}
-
-async function cancelStatement(handle: string): Promise<boolean> {
-  if (!IS_SPCS || !handle || handle.startsWith('local_')) return false;
-  try {
-    const token = getSpcsToken();
-    const headers: Record<string, string> = {
-      'Authorization': `Bearer ${token}`, 'Content-Type': 'application/json',
-      'X-Snowflake-Authorization-Token-Type': 'OAUTH',
-    };
-    const r = await fetch(`https://${SNOWFLAKE_HOST}/api/v2/statements/${handle}/cancel`, { method: 'POST', headers });
-    return r.ok;
-  } catch { return false; }
-}
 
 app.get('/api/matrix/road-filter-available', async (_req, res) => {
   try {
